requirements.txt
annotated-types==0.7.0
anyio==4.5.2
click==8.1.8
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.16.1
fsspec==2025.3.0
h11==0.14.0
idna==3.10
jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
pillow==10.4.0
pydantic==2.10.6
pydantic-core==2.27.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.44.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing-extensions==4.13.0
uvicorn==0.33.0
